api/requirements.txt
annotated-types==0.7.0
azure-core==1.37.0
azure-cosmos==4.14.3
azure-functions==1.24.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
markupsafe==3.0.3
pydantic==2.12.5
pydantic-core==2.41.5
requests==2.32.5
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
werkzeug==3.1.4
